Added Enemys

+++ src/main/java/game/mobs/Enemy.java
@@ -1,14 +1,14 @@
 package game.mobs;
 
 public class Enemy extends Entity{
-    private int speed;
+    private double speed;
     private int reward;
-    public Enemy(String name, int maxHealth, int atk, double atkSpeed, double range, Element element, int speed, int reward){
+    public Enemy(String name, int maxHealth, int atk, double atkSpeed, double range, Element element, double speed, int reward){
         super(name, maxHealth, atk, atkSpeed, range, element);
         this.speed = speed;
         this.reward = reward;
     }
-    public int getSpeed() {
+    public double getSpeed() {
         return speed;
     }
     public int getReward() {
